feat(protocol,crypto): graduate the verification-verdict vocabulary to @motebit/protocol
The structured verdict vocabulary (IntegrityVerdict through
VerificationVerdict, RepairInstruction, VerdictSubject) moves to
packages/protocol/src/verification-verdict.ts — the closed verdict
vocabulary's home, and the prerequisite for EvalAttestation: the
attestation embeds whole per-axis verdicts per measurement, and
@motebit/wire-schemas (which may only depend on protocol) needs the
shape for its zod parity block.

Crypto re-exports every name type-only (the EvidenceRef graduation
precedent; zero runtime monorepo deps preserved), so its public surface
is name-identical — verified via api-extractor baselines regenerated for
both packages. VerdictSubject is RESTATED in protocol as an explicit
closed literal union (deriving from crypto's VerifyResult would be a
cycle); crypto carries a compile-time ArtifactType-extends-VerdictSubject
assert as the drift lock. Widened additively with the Auditor's four
measurement subjects: succession, revocation, bond_commitment,
solvency_proof (no exhaustive switches over the union exist outside
crypto/verifier — verified in the pre-implementation review).

No behavior change: protocol 446, crypto 691, verifier 52,
state-export-client 91 tests green; full-repo typecheck green.
