+++ packages/crypto/src/index.ts
@@ -41,6 +41,16 @@ import type {
   DigestRef,
   DigestAlgorithm,
   ProjectionClass,
+  IntegrityVerdict,
+  IdentityBindingVerdict,
+  AuthorityVerdict,
+  RevocationStatus,
+  TemporalBasis,
+  RevocationFreshness,
+  RevocationVerdict,
+  RepairInstruction,
+  VerdictSubject,
+  VerificationVerdict,
 } from "@motebit/protocol";
 import { hash, isScopeNarrowed } from "./signing.js";
 // The @noble/ed25519 SHA-512 binding is performed in suite-dispatch.ts
@@ -392,73 +402,34 @@ export type ArtifactType = VerifyResult["type"];
 // #2 — revocation as a freshness BASIS (not a bare label), `repair` first-class.
 // ===========================================================================
 
-/** Did the signature verify over canonical bytes? The one clockless, always-establishable axis. */
-export type IntegrityVerdict = "verified" | "invalid";
+// The verdict vocabulary (IntegrityVerdict … VerificationVerdict) graduated
+// to @motebit/protocol — the closed verdict vocabulary's home — in the
+// 2026-07-08 EvalAttestation arc: EvalAttestation embeds whole verdicts per
+// measurement, and @motebit/wire-schemas (which may only see protocol) needs
+// the shape for its zod parity block. Re-exported here type-only (crypto
+// keeps zero runtime monorepo deps; the EvidenceRef graduation precedent) so
+// the verify family's public surface is unchanged.
+export type {
+  IntegrityVerdict,
+  IdentityBindingVerdict,
+  AuthorityVerdict,
+  RevocationStatus,
+  TemporalBasis,
+  RevocationFreshness,
+  RevocationVerdict,
+  RepairInstruction,
+  VerdictSubject,
+  VerificationVerdict,
+};
 
-/**
- * The identity-binding rung (docs/doctrine/identity-binding-verification.md):
- * how strongly the signing key is bound to the claimed `motebit_id`. `sovereign`
- * = the id commits to the genesis key (offline, no operator); `anchored` =
- * transparency-log inclusion confirmed on-chain; `pinned` = time-valid in the
- * operator's succession chain; `unverified` = integrity only, key→id NOT
- * established (the embedded-key footgun, named honestly); `invalid` = a binding
- * was claimed and failed.
- */
-export type IdentityBindingVerdict = "sovereign" | "anchored" | "pinned" | "unverified" | "invalid";
-
-/**
- * Whether the authority (delegation grant / token scope) covering the action
- * holds. `valid` = a check ran and held; `expired` / `not_yet_valid` = outside
- * the validity window (past / future) under the verdict's `temporalBasis`;
- * `insufficient` = scope/structure does not confer it; `unknown` = no authority
- * context to evaluate (a bare receipt) — never a manufactured pass.
- */
-export type AuthorityVerdict = "valid" | "expired" | "not_yet_valid" | "insufficient" | "unknown";
-
-/** Revocation status. `unchecked` NEVER reads as "not revoked" — it is its own axis value. */
-export type RevocationStatus = "fresh" | "stale" | "unchecked" | "revoked";
-
-/**
- * The temporal basis a time-dependent axis was evaluated against. `clockless` =
- * no clock needed; `local_clock` = the verifier's own clock (first-person time,
- * self-borne risk); `ledger_anchored` = a chain slot/height — ORDERING, not
- * wall-clock (see docs/doctrine/verify-family-fail-closed.md).
- */
-export type TemporalBasis = "clockless" | "local_clock" | "ledger_anchored";
-
-/**
- * How a `RevocationStatus` was established, so an OFFLINE/P2P verifier can dial
- * its OWN staleness tolerance rather than accept a bare "stale" (co-designed
- * with consumer #2). `basis` is an agility axis — append a new value when a
- * consumer needs one (docs/doctrine/agility-as-role.md). `asOf` carries the
- * revocation set's wall-clock timestamp AND, when available, its deterministic
- * chain anchor (slot/height) — branch on the anchor, since chain time is
- * ordering, not wall-clock.
- *
- * `basis` is an evidence-grade ladder, weakest to strongest: `asserted` <
- * `stapled` < `ledger`. `asserted` = holder-asserted freshness with NO external
- * anchor ("we hold this set, refreshed at T, on our own say-so") — a consumer
- * SHOULD down-weight it relative to a `stapled` signed freshness proof or a
- * `ledger` chain root. The verdict carries the basis; it does NOT assign a
- * weight — the consumer holds the tolerance and sets where its acceptance line
- * falls. When `basis` is `asserted` the `asOf.anchor` is absent and the
- * `asOf.timestamp_ms` is a holder-asserted bound, not an orderable anchor.
- */
-export interface RevocationFreshness {
-  basis: "asserted" | "stapled" | "ledger";
-  asOf: {
-    /** Wall-clock ms the revocation set was current as of, when known. */
-    timestamp_ms?: number;
-    /** Deterministic chain anchor the set was current as of, when known. */
-    anchor?: { chain: string; slot?: number; height?: number };
-  };
-}
-
-export interface RevocationVerdict {
-  status: RevocationStatus;
-  /** Present when `status` derives from a freshness basis (a stapled proof or a ledger root). */
-  freshness?: RevocationFreshness;
-}
+// Compile-time drift lock for the graduation: protocol RESTATES VerdictSubject
+// as an explicit closed literal union (it cannot derive from crypto's
+// VerifyResult — that would be a dependency cycle), so containment is asserted
+// here instead. A new VerifyResult member that protocol's union doesn't carry
+// turns this line into a build error rather than a silent divergence.
+type _ArtifactTypeWithinVerdictSubject = ArtifactType extends VerdictSubject ? true : never;
+const _assertVerdictSubjectCoverage: _ArtifactTypeWithinVerdictSubject = true;
+void _assertVerdictSubjectCoverage;
 
 /** A reference to the evidence an axis was established from (a receipt hash, a key id, a revocation root). */
 // EvidenceRef + the evidence-provenance vocabulary graduated to @motebit/protocol
@@ -469,55 +440,9 @@ export interface RevocationVerdict {
 // `verifyEvidenceProvenance` below (evidence-provenance arc).
 export type { EvidenceRef, EvidenceProvenance, DigestRef, DigestAlgorithm, ProjectionClass };
 
-/**
- * Machine-readable repair instruction for a failing axis — first-class, not
- * optional-if-time (consumer #2). `code` is for programmatic branching; `fix` +
- * `canonical` are the legibility-on-contact pair ("learn the one axis you hit,"
- * not the whole verifier). Same shape-of-intent as the gate-repair contract
- * (scripts/lib/gate-report.ts), applied to verification.
- */
-export interface RepairInstruction {
-  /** Stable, machine-readable reason code (e.g. "revocation.unchecked", "identity.embedded_key_only"). */
-  code: string;
-  /** Which axis failed. */
-  axis: "integrity" | "identityBinding" | "authority" | "revocation";
-  /** One-line human summary of what's wrong. */
-  summary: string;
-  /** The canonical source of truth to consult or fix, when applicable. */
-  canonical?: string;
-  /** The concrete next step to establish the axis. */
-  fix: string;
-}
-
-/**
- * The structured verification verdict. DELIBERATELY carries no top-level `valid`
- * boolean: a consumer MUST branch on the axis it depends on, so an `unchecked` /
- * `stale` / `unverified` / `unknown` result cannot silently read as a pass.
- * `repair` is present whenever any axis is not in its passing state
- * (`integrity: "verified"`, `identityBinding` of sovereign/anchored/pinned,
- * `authority: "valid"`, `revocation.status: "fresh"`); its presence-on-failure
- * is enforced by the conformance corpus, not left to the implementer.
- */
-/**
- * What a `VerificationVerdict` is about. A superset of `ArtifactType` — adds
- * `delegation_token` for the per-tick token/grant path (which is not a
- * top-level auto-detected artifact). Widening is additive: existing values are
- * unchanged.
- */
-export type VerdictSubject = ArtifactType | "delegation_token";
-
-export interface VerificationVerdict {
-  /** What the verdict is about. */
-  type: VerdictSubject;
-  integrity: IntegrityVerdict;
-  identityBinding: IdentityBindingVerdict;
-  authority: AuthorityVerdict;
-  revocation: RevocationVerdict;
-  temporalBasis: TemporalBasis;
-  evidenceBasis: readonly EvidenceRef[];
-  /** Present whenever any axis is not passing; absent only on a clean pass. */
-  repair?: RepairInstruction;
-}
+// RepairInstruction, VerdictSubject, and VerificationVerdict graduated to
+// @motebit/protocol with the rest of the verdict vocabulary (see the
+// re-export block above).
 
 export interface VerifyOptions {
   expectedType?: ArtifactType;
